Trying to make it so that I can 'take' the tags that are specified in the url and turn them into selected. Difficult because I have three or something tags types in tagStore. Trying to implement the logic in the watchStore. unsure how I want to intially fetch the posts on the index page, AND unsure who should fetch new posts when local url is updated
